requirements.txt
asgiref==3.8.1
black==24.3.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
filelock==3.13.4
identify==2.5.35
mypy-extensions==1.0.0
mysqlclient==2.2.4
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.7.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.4.4
virtualenv==20.25.1
